==> GCloud/01_Regression/01.0 - Regression Data Generation.ipynb <==
import numpy as np
import pandas as pd
from sklearn import *
import matplotlib.pyplot as plt
%matplotlib inline

sample_size = 5000


data1,target1 = datasets.make_circles(n_samples=sample_size, factor=.1, noise=0.2)
target1 = (3*data1[:,0])-(16*data1[:,1]) + (0.5*data1[:,0]*data1[:,1]) + np.random.normal(0,2,size=sample_size)


data2,target2 = datasets.make_circles(n_samples=sample_size, factor=.5, noise=0.2)
target2 = np.power(data2[:,0],2) + 10*np.power(data2[:,1],3) + (50*data2[:,0]*np.power(data2[:,1],2)) + np.random.normal(0,2,size=sample_size)

data3,target3 = datasets.make_moons(n_samples=sample_size,noise=0.2)
data3[:,0] = (2 * (data3[:, 0]-(-1))/(3))-1
data3[:,1] = (2 * (data3[:, 1]-(-1))/(2))-1
target3 = (50*data3[:,0]*np.sin(data3[:,1])) + (50*data3[:,1]*np.cos(data3[:,0]))

data4,target4 = datasets.make_moons(n_samples=sample_size,noise=0.2)

temp = np.copy(data4[:, 0])
data4[:, 0] = data4[:, 1]
data4[:, 1] = temp
data4[:,0] = (2 * (data4[:, 0]-(-1))/(2))-1
data4[:,1] = (2 * (data4[:, 1]-(-1))/(3))-1

target4 = (30*data1[:,0])-(16*data1[:,1]) - (1.5*data1[:,0]*data1[:,1]) + np.random.normal(0,1,size=sample_size)

data = np.concatenate((data1, data2, data3, data4), axis=0)
target = np.concatenate((target1,target2,target3,target4),axis=0)
alpha = np.concatenate((np.zeros(sample_size),np.ones(sample_size),np.zeros(sample_size),np.ones(sample_size)), axis=0)
beta = np.concatenate((np.zeros(sample_size),np.zeros(sample_size),np.ones(sample_size),np.ones(sample_size)), axis=0)

data_frame = pd.DataFrame(data = data,columns=["x","y"])
data_frame["alpha"] = pd.Series(alpha).map(lambda v: 'ax01' if v==0 else 'ax02')
data_frame["beta"] = pd.Series(beta).map(lambda v: 'bx01' if v==0 else 'bx02')
data_frame["target"] = target

data_frame.describe()

distribution =  ([0] * sample_size) + ([1] * sample_size) + ([2] * sample_size) + ([3] * sample_size)

splitter =  model_selection.StratifiedShuffleSplit(n_splits=1, test_size=0.25, random_state=0)
splits = list(splitter.split(X=data_frame.iloc[:,[0,1,2,3]],y=distribution))
learn_index = splits[0][0]
test_index = splits[0][1]

learn_df = data_frame.iloc[learn_index,:]

size2 = int(len(learn_df)/4)
distribution2 =  ([0] * size2) + ([1] * size2) + ([2] * size2) + ([3] * size2)


splitter =  model_selection.StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=0)
splits = list(splitter.split(X=learn_df.iloc[:,[0,1,2,3]],y=distribution2))
train_index = splits[0][0]
valid_index = splits[0][1]


train_df = learn_df.iloc[train_index,:]
print(len(train_df))

valid_df = learn_df.iloc[valid_index,:]
print(len(valid_df))

test_df = data_frame.iloc[test_index,:]
print(len(test_df))


train_df.to_csv(path_or_buf="data/train-data.csv", header=False, index=True)
valid_df.to_csv(path_or_buf="data/valid-data.csv", header=False, index=True)
test_df.to_csv(path_or_buf="data/test-data.csv", header=False, index=True)
